+++ packages/vike/node/runtime/html/serializeContext.ts
@@ -18,7 +18,7 @@ import { getPropKeys, getPropVal, setPropVal } from './propKeys.js'
 import type { GlobalContextServerInternal } from '../globalContext.js'
 import type { PageContextCreated } from '../renderPage/createPageContextServerSide.js'
 import type { PageContextBegin } from '../renderPage.js'
-import type { PageContextCspNonce } from './injectAssets/inferHtmlTags.js'
+import type { PageContextCspNonce } from '../renderPage/csp.js'
 
 const passToClientBuiltInPageContext = [
   'abortReason',
